refactor: 불필요한 코드 제거 및 databinding으로 변경
