+++ presentation/build.gradle.kts
@@ -19,7 +19,7 @@ android {
     }
 
     buildFeatures {
-        viewBinding = true
+        dataBinding = true
     }
 }
